Fixed Issue with 2Clock Type not showing Dreiviertel. Extracted Telnet-func to external File.
